Handled driver path location

+++ src/main/java/base/TestBase.java
@@ -31,23 +31,25 @@ public class TestBase {
 		}
 	}
     public void setDriverPath() {
-        if (PlatformUtil.isMac()) {
-            System.setProperty("webdriver.chrome.driver", "chromedriver");
-        }
-        if (PlatformUtil.isWindows()) {
-            System.setProperty("webdriver.chrome.driver", System.getProperty("user.dir")+"/chromedriver.exe");
+		if(driver==null) {
+			if (PlatformUtil.isMac()) {
+				System.setProperty("webdriver.chrome.driver", System.getProperty("user.dir") + "/chromedriver");
+			}
+			if (PlatformUtil.isWindows()) {
+				System.setProperty("webdriver.chrome.driver", System.getProperty("user.dir") + "/chromedriver.exe");
 
-        }
-        if (PlatformUtil.isLinux()) {
-            System.setProperty("webdriver.chrome.driver", "chromedriver_linux");
-        }
-        driver =new ChromeDriver();
-        driver.manage().window().maximize();
-        driver.manage().deleteAllCookies();
-        driver.manage().timeouts().pageLoadTimeout(TestUtils.PAGE_LOAD_TIMEOUT,TimeUnit.SECONDS);
-        driver.manage().timeouts().implicitlyWait(TestUtils.IMPLICIT_WAIT, TimeUnit.SECONDS);
+			}
+			if (PlatformUtil.isLinux()) {
+				System.setProperty("webdriver.chrome.driver", System.getProperty("user.dir") + "/chromedriver_linux");
+			}
+			driver = new ChromeDriver();
+		 }
+			driver.manage().window().maximize();
+			driver.manage().deleteAllCookies();
+			driver.manage().timeouts().pageLoadTimeout(TestUtils.PAGE_LOAD_TIMEOUT,TimeUnit.SECONDS);
+			driver.manage().timeouts().implicitlyWait(TestUtils.IMPLICIT_WAIT, TimeUnit.SECONDS);
     }
-	    public static void navigateToUrl() {
+	    public void navigateToUrl() {
 			driver.get(prop.getProperty("url"));
 	    }
